Add files via upload
The dataset utilised for this report is PAMAP2, an activity monitoring dataset that contains data from 18 distinct physical activities (such as walking, cycling, ironing, etc) completed by 9 different subjects — eight men and one woman — wearing 3 inertial measurement units and a heart rate monitor.

This report will leverage PAMAP2 dataset to derive practical knowledge that can be used for activity recognition and intensity estimation, while developing and applying algorithms for data processing, feature extraction and classification. It would also provide insights that can enable the creation of a software and/or hardware for fitness tracking. This project has three objectives, which are as follows:

1) Conduct in-depth exploratory data analysis and manage missing or contaminated data correctly.

2) Develop and test at least one hypothesis for a relationship between a single pair of attributes.

3) Develop and test at least one model which uses multiple attributes to make predictions.